[refactor] : router 객체 분리 및 게시판 컴포넌트 분리
