+++ src/pages/community/index.ts
@@ -1,4 +1,5 @@
-export { default as Board } from './Board';
+export { default as NoticeBoard } from './NoticeBoard';
+export { default as FreeBoard } from './FreeBoard';
 export { default as BoardWrite } from './BoardWrite';
 export { default as BoardView } from './BoardView';
 export { default as BoardModify } from './BoardModify';
